+++ src/app/(home)/cart/page.tsx
@@ -57,6 +57,10 @@ export const metadata: Metadata = {
     index: false, // ✅ best practice: prevent indexing cart pages
     follow: true,
     nocache: true,
+    googleBot: {
+      index: false,
+      follow: false,
+    },
   },
 };
