+++ src/app/(store)/layout.tsx
@@ -2,15 +2,16 @@ import { StoreHeader } from "@/components/layout/store/nav";
 import "../globals.css";
 import type { Metadata } from "next";
 import { Fragment } from "react";
-import { siteConfig } from "@/config/site";
 
-// Metadata for the application
 export const metadata: Metadata = {
-  title: siteConfig.siteTitle,
-  description: siteConfig.description,
-  keywords: siteConfig.keywords,
-  icons: {
-    icon: "/favicon.ico",
+  robots: {
+    index: false, // ✅ best practice: prevent indexing cart pages
+    follow: true,
+    nocache: true,
+    googleBot: {
+      index: false,
+      follow: false,
+    },
   },
 };
