+++ src/modules/skeletons/payment-success-skeleton.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Skeleton } from "@/components/ui/skeleton";
 
 export default function PaymentSuccessSkeleton() {
